[IMP] web: add discuss type def generation to pre commit hook
Discuss uses a script to generate type definitions for its models.
However, running it on each PR is cumbersome, which can lead to
outdated type definitions.

This commit adds the script to the web pre-commit hook. The script
runs only when Discuss related modules are modified.
