+++ addons/im_livechat/static/src/core/public_web/@types/models.d.ts
@@ -1,11 +1,4 @@
 declare module "models" {
-    export interface DiscussApp {
-        defaultLivechatCategory: DiscussAppCategory;
-        livechats: DiscussChannel[];
-    }
-    export interface DiscussAppCategory {
-        livechat_channel_id: LivechatChannel;
-    }
     export interface DiscussChannel {
         appAsLivechats: DiscussApp;
     }
