+++ addons/im_livechat/static/src/core/public_web/discuss_app/@types/models.d.ts
@@ -1,0 +1,9 @@
+declare module "models" {
+    export interface DiscussApp {
+        defaultLivechatCategory: DiscussAppCategory;
+        livechats: DiscussChannel[];
+    }
+    export interface DiscussAppCategory {
+        livechat_channel_id: LivechatChannel;
+    }
+}
