+++ addons/mail/static/src/discuss/core/common/@types/models.d.ts
@@ -9,6 +9,7 @@ declare module "models" {
         channelMembers: ChannelMember[];
     }
     export interface Message {
+        channel_id: DiscussChannel;
         channelMemberHaveSeen: Readonly<ChannelMember[]>;
         hasEveryoneSeen: boolean|undefined;
         hasNewMessageSeparator: boolean;
@@ -36,6 +37,7 @@ declare module "models" {
     }
     export interface Thread {
         _computeOfflineMembers: () => ChannelMember[];
+        allow_invite_by_email: Readonly<boolean>;
         allowCalls: Readonly<boolean>;
         allowDescription: Readonly<boolean>;
         allowedToLeaveChannelTypes: Readonly<string[]>;
@@ -61,7 +63,6 @@ declare module "models" {
         firstUnreadMessage: Message;
         group_ids: ResGroups[];
         hasMemberList: Readonly<boolean>;
-        hasOtherMembersTyping: boolean;
         hasSeenFeature: boolean;
         hasSelfAsMember: Readonly<boolean>;
         invitationLink: Readonly<unknown|string>;
@@ -72,6 +73,7 @@ declare module "models" {
         lastMessageSeenByAllId: undefined|number;
         lastSelfMessageSeenByEveryone: Message;
         leaveChannel: () => Promise<void>;
+        leaveChannelRpc: () => void;
         markAsFetched: () => Promise<void>;
         markedAsUnread: boolean;
         markingAsRead: boolean;
@@ -83,7 +85,6 @@ declare module "models" {
         notifyDescriptionToServer: (description: unknown) => Promise<unknown>;
         offlineMembers: ChannelMember[];
         onlineMembers: ChannelMember[];
-        otherTypingMembers: ChannelMember[];
         rename: (name: string) => Promise<void>;
         scrollUnread: boolean;
         self_member_id: ChannelMember;
@@ -91,7 +92,6 @@ declare module "models" {
         showUnreadBanner: Readonly<boolean>;
         toggleBusSubscription: boolean;
         typesAllowingCalls: Readonly<string[]>;
-        typingMembers: ChannelMember[];
         unknownMembersCount: Readonly<number>;
     }
